mktree: More generalized banding
Previously, the banding algorithm depended on understanding call
sequences.  This was necessary because it otherwise only accounted for
the innermost frame of each SchedNode, so the call sequences were
necessary to disambiguate the different call sequence regions.  This
patch changes the banding algorithm to use the whole call stack of
each schedule node.  Now it only fails for honest loops in the code.

This also moves more logic into TreeNode, gives TreeNodes explicit
types (instead of deducing this from the type of TreeNode.val), and
records user-friendly region labels (though they're currently unused).