tool: Rework Analyzer flow and add much more detail
This describes the Analyzer algorithm in depth and with code listings
and explains how it corresponds to SIM commutativity.  It first
attacks it as a concrete algorithm, so that expanding it to symbolic
execution is a natural extension.  It then goes into a detailed
example of how it analyzes rename/rename before giving its
commutativity condition.