Revert "bench: Make CPU 0 perform both setup operations and main operations"
This reverts commit d531b636aa22a5bf7432f234080e59b3cea4fed4.

For whatever reason, this caused the Intel branch predictor's brain to
explode and desynchronized the threads by TENS OF THOUSANDS of cycles
while also increasing overall CPI by a factor of four.  I've tried
fixing it, but it seems like no matter what I do, the moment it
captures the phase from the TSC, the branch predictor explodes.