bench: Fix handling of IQR when bucket count exceeds limit
Previously, the IQR computation assumed it the histogram data was
continuous, which was fine when the buckets spanned more than one
value, but broke down when each bucket spanned less than one value.
Make the IQR itself discrete.  If the KDE bandwidth is zero, use a
delta function for the kernel.