Rewrite \code for the umpteenth time
The old definition didn't work in \underbrace of all things.  Finally
scrap the use of \lstinline entirely.  The only tricky part is that we
use literal _'s in lots of code arguments, so we have to typeset those
correctly.

This also means I no longer need a funny, different definition of
\cmd.