scalability: Move back before rule chapter for bottom-up flow
We use this like a lemma in the rule, so it makes more sense to put it
before the rule chapter.  This way, the rule is also stated at the end
of these two chapters, rather than in the middle.

Changed the titles back to their original titles, since those fit this
bottom-up flow better.  The has the added advantage that I now have a
chapter named after the rule again.